% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportReconciledGeographies.R
\name{getCentroid}
\alias{getCentroid}
\title{Get the centroid from two geograhies.}
\usage{
getCentroid(geoA, geoB)
}
\arguments{
\item{geoA}{The first geography.}

\item{geoB}{The second geography.}
}
\value{
A character vector.
}
\description{
Get the centroid from two geograhies.
}
\examples{
getCentroid(
 polyA[res_to_print$unigeokey_A[i],],
 polyB[res_to_print$unigeokey_B[i],]
)
}
