% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileGeographies.R
\name{getUniqueReconciliationKey}
\alias{getUniqueReconciliationKey}
\title{Get a unique reconciliation key based on results from reconcileGeographies().}
\usage{
getUniqueReconciliationKey(res)
}
\arguments{
\item{res}{Results from reconcileGeographies().}
}
\value{
A data.frame.
}
\description{
Get a unique reconciliation key based on results from reconcileGeographies().
}
