% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poly_a}
\alias{poly_a}
\title{Polygons of the Mexican state of Aguascalientes (2010).}
\format{589 simple features with corresponding data:
\describe{
  \item{gid}{unique id}
  \item{geometry}{geometry of the feature}
  \item{population}{randomly generated}
  \item{party_a_votes}{randomly generated}
  \item{party_b_votes}{randomly generated}
}}
\source{
\url{http://gaia.inegi.org.mx/}
}
\usage{
poly_a
}
\description{
Simple Features containing randomly generated population and votes variables.
diamonds.
}
\keyword{datasets}
