% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poly_b}
\alias{poly_b}
\title{Polygons of the Mexican state of Aguascalientes (2017).}
\format{602 simple features with corresponding data:
\describe{
  \item{gid}{unique id}
  \item{geometry}{geometry of the feature}
  \item{population}{randomly generated}
  \item{party_a_votes}{randomly generated}
  \item{party_b_votes}{randomly generated}
}}
\source{
\url{http://gaia.inegi.org.mx/}
}
\usage{
poly_b
}
\description{
Simple Features containing randomly generated population and votes variables.
diamonds.
}
\keyword{datasets}
