% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileGeographies.R
\name{reconcileData}
\alias{reconcileData}
\title{Reconcile data based on results from reconcileGeographies().}
\usage{
reconcileData(res, dataA, dataB, idA = NULL, idB = NULL, varA, varB,
  return_spatial = "A")
}
\arguments{
\item{res}{Results from reconcileGeographies().}

\item{dataA}{The first data object. Either a spatial object with a data.frame or a data.frame.}

\item{dataB}{The second data object. Either a spatial object with a data.frame or a data.frame.}

\item{idA}{The name of the id column in the first object. If not provided, the first column is assume to be the ID.}

\item{idB}{The name of the id column in the second object. If not provided, the first column is assume to be the ID.}

\item{varA}{An ordered character vector with the names of the variables to concile from the first spatial object.}

\item{varB}{An ordered character vector with the names of the variables to concile from the second spatial object.}

\item{return_spatial}{Whether to return a spatial object (set to "A" or "B") or a simple data.frame (set to FALSE). "A" or "B" determines which geometries to return, either from the first or the seconf spatial object.}
}
\value{
A Simple Feature with a data.frame or only a data.frame.
}
\description{
Reconcile data based on results from reconcileGeographies().
}
\examples{

}
