% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileGeographies.R
\name{reconcileGeographies}
\alias{reconcileGeographies}
\title{Reconcile geographies (features) from two spatial polygon objects.}
\usage{
reconcileGeographies(polyA, polyB, idA = NULL, idB = NULL,
  project_crs = NULL, dist_buffer = 5)
}
\arguments{
\item{polyA}{The first spatial polygon object.}

\item{polyB}{The second spatial polygon object.}

\item{idA}{The name of the id column in the first object. If not provided, the first column is assume to be the ID.}

\item{idB}{The name of the id column in the second object. If not provided, the first column is assume to be the ID.}

\item{project_crs}{The EPSG coordinate system. The unit must be the metre.}

\item{dist_buffer}{The distance for the buffer. Default is 5 metres.}
}
\value{
A data.frame relating the geographies in the two spatial objects (one-to-one, one-to-many) or many-to-many.
data(polygons, package = "recogeo")
res <- reconcileGeographies(poly_a, poly_b)
}
\description{
Reconcile geographies (features) from two spatial polygon objects.
}
