% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportReconciledGeographies.R
\name{reportReconciledGeographies}
\alias{reportReconciledGeographies}
\title{Render a report comparing the reconciled geometries.}
\usage{
reportReconciledGeographies(res, polyA, polyB, output = "html",
  type = c("intersects", "contains"))
}
\arguments{
\item{res}{The result from reconcileGeometries().}

\item{polyA}{The first spatial polygon object passed to reconcileGeometries().}

\item{polyB}{The second spatial polygon object passed to reconcileGeometries().}

\item{output}{The type of document to produce: either 'html', 'pdf' or 'word'.}

\item{type}{A character vector with one or more type of reconciliations to print out: 'intersects', 'contains' and 'same'.}
}
\value{
A data.frame relating the geographies in the two spatial objects (one-to-one, one-to-many) or many-to-many.
}
\description{
Render a report comparing the reconciled geometries.
}
\examples{
\dontrun{reportReconciledGeographies(res, polyA, polyB)}
}
