% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileGeographies.R
\name{testIntersectionArea}
\alias{testIntersectionArea}
\title{Test that the intersection area of two geometries is larger than a minimum}
\usage{
testIntersectionArea(unigeokey_A, unigeokey_B, polyA, polyB,
  min_inters_area)
}
\arguments{
\item{unigeokey_A}{The unique ID of the geometry in the first spatial object.}

\item{unigeokey_B}{The unique ID of the geometry in the second spatial object.}

\item{polyA}{The first spatial object.}

\item{polyB}{The second spatial object.}

\item{min_inters_area}{The minimum intersecting area.}
}
\value{
A logical vector.
}
\description{
Test that the intersection area of two geometries is larger than a minimum
}
\examples{
mapply(testIntersectionArea, unigeokeys_A, unigeokeys_B, MoreArgs = list(polyA, polyB, min_inters_area))
}
