% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportReconciledGeographies.R
\name{testReconciledGeographies}
\alias{testReconciledGeographies}
\title{Test the spatial charateristics of reconciled geographies.}
\usage{
testReconciledGeographies(res, polyA, polyB, idA = NULL, idB = NULL,
  project_crs = NULL)
}
\arguments{
\item{res}{The result from reconcileGeometries().}

\item{polyA}{The first spatial polygon object passed to reconcileGeometries().}

\item{polyB}{The second spatial polygon object passed to reconcileGeometries().}

\item{idA}{The name of the id column in the first object. If not provided, the first column is assume to be the ID.}

\item{idB}{The name of the id column in the second object. If not provided, the first column is assume to be the ID.}

\item{project_crs}{The EPSG coordinate system. The unit must be the metre.}
}
\value{
A data.frame.
}
\description{
Test the spatial charateristics of reconciled geographies.
}
\examples{
\dontrun{testReconciledGeographies(res, polyA, polyB)}
}
