% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportReconciledGeographies.R
\name{typesetFigure}
\alias{typesetFigure}
\title{Typeset the figure.}
\usage{
typesetFigure(i, coords, Ai, Bi, type)
}
\arguments{
\item{i}{The number of figure.}

\item{coords}{The geographic coordinates of the figure.}

\item{Ai}{The index of the geography in the first spatial object.}

\item{Bi}{The index of the geography in the second spatial object.}

\item{type}{The type of reconciliation.}
}
\value{
A character vector.
}
\description{
Typeset the figure.
}
\examples{
\dontrun{
typesetFigure(i, getCentroid(
 polyA[res_to_print$unigeokey_A[i],],
 polyB[res_to_print$unigeokey_B[i],
)
}
}
